<fix>: <컨벤션 유지를 위한 디렉토리 수정 + 프로필 디버깅>
- 컨벤션 유지를 위해 디렉토리 구조와 파일명을 수정했습니다.
- 프로필 페이지를 메인페이지에서 넘어갈 때, navigate로 state를 넘겨주지 않아
- null이 되는 상황에서 에러가 발생해 해당 부분을 디버깅하는 코드를 추가하였습니다.
