+++ .env
@@ -1,4 +1,10 @@
 VITE_APP_LOCAL_HOST=http://localhost:5173/
 VITE_APP_PRODUCTION_URL=https://study-badge.vercel.app/
+VITE_APP_KAKAO_REDIRECT_URL_DEV=http://localhost:5173/auth/kakao/callback
+VITE_APP_KAKAO_REDIRECT_URL_PROD=https://study-badge.vercel.app/auth/kakao/callback
+VITE_APP_NAVER_REDIRECT_URL_DEV=http://localhost:5173/auth/naver/callback
+VITE_APP_NAVER_REDIRECT_URL_PROD=https://study-badge.vercel.app/auth/naver/callback
 
+VITE_APP_KAKAO_LOGIN_CLIENT_KEY=224eb50d7feba5e817fcf2ff83d6187c
+VITE_APP_NAVER_LOGIN_CLIENT_KEY=_Pm6GirABhVnjO1SgAze
 VITE_APP_TOSS_TEST_CLIENT_KEY=test_gck_docs_Ovk5rk1EwkEbP0W43n07xlzm
