+++ src/pages/CreateStudyPage.tsx
@@ -1,0 +1,13 @@
+import CreateStudy from "../components/main/CreateStudy";
+
+const CreateStudyPage = (): JSX.Element => {
+  // 함수 작성 금지
+
+  return (
+    <>
+      <CreateStudy />
+    </>
+  );
+};
+
+export default CreateStudyPage;
