+++ src/pages/Login.tsx
@@ -1,4 +1,4 @@
-import LoginUser from "../components/layout/login/LoginUser";
+import LoginUser from "../components/auth/LoginUser";
 
 const Login = (): JSX.Element => {
   // 함수 작성 금지
