+++ src/pages/PasswordResetPage.tsx
@@ -1,0 +1,11 @@
+import PasswordReset from "../components/auth/PasswordReset";
+
+const PasswordResetPage = (): JSX.Element => {
+  return (
+    <>
+      <PasswordReset />
+    </>
+  );
+};
+
+export default PasswordResetPage;
