+++ src/pages/SignUp.tsx
@@ -1,4 +1,4 @@
-import SignUpUser from "../components/layout/login/SignUpUser";
+import SignUpUser from "../components/auth/SignUpUser";
 
 const SignUp = (): JSX.Element => {
   // 함수 작성 금지
