+++ src/router/Router.tsx
@@ -4,16 +4,16 @@ import Error from "../pages/Error";
 import ChannelPage from "../pages/ChannelPage";
 import Login from "../pages/Login";
 import SignUp from "../pages/SignUp";
-import PasswordReset from "../pages/PasswordReset";
-import SocialLoginCallback from "../components/layout/login/SocialLoginCallback";
 import ProfilePage from "../pages/ProfilePage";
 import PaymentListPage from "../pages/PaymentListPage";
 import Success from "../components/payment/Success";
 import Fail from "../components/payment/Fail";
-import CreateStudy from "../pages/CreateStudy";
 import Profile from "../components/profile/Profile";
 import ChannelBook from "../components/channel/ChannelBook";
 import ScheduleEdit from "../components/schedule/ScheduleEdit";
+import CreateStudyPage from "../pages/CreateStudyPage";
+import SocialLoginCallback from "../components/auth/SocialLoginCallback";
+import PasswordResetPage from "../pages/PasswordResetPage";
 
 const Router = (): JSX.Element => {
   return (
@@ -23,8 +23,9 @@ const Router = (): JSX.Element => {
       <Route path="/:type/:status/:category/:keywordValue/:order/:page" element={<Main />} />
       <Route path="/login" element={<Login />} />
       <Route path="/SignUp" element={<SignUp />} />
-      <Route path="/PasswordReset" element={<PasswordReset />} />
-      <Route path="/SocialLoginCallback" element={<SocialLoginCallback />} />
+      <Route path="/PasswordReset" element={<PasswordResetPage />} />
+      <Route path="/auth/kakao/callback" element={<SocialLoginCallback />} />
+      <Route path="/auth/naver/callback" element={<SocialLoginCallback />} />
       <Route path="/profile" element={<ProfilePage />}>
         <Route path="/profile/myInfo" element={<Profile />} />
         <Route path="/profile/paymentList" element={<PaymentListPage />} />
@@ -37,7 +38,7 @@ const Router = (): JSX.Element => {
         <Route path="/channel/:channelId/information/information_edit" element={<ChannelBook />} />
         <Route path="/channel/:channelId/schedule/schedule_edit" element={<ScheduleEdit />} />
       </Route>
-      <Route path="/createStudy" element={<CreateStudy />} />
+      <Route path="/createStudy" element={<CreateStudyPage />} />
     </Routes>
   );
 };
