+++ src/services/auth.ts
@@ -25,10 +25,14 @@ export const initiateSocialLogin = (provider: "naver" | "kakao") => {
   window.location.href = `/api/auth/${provider}`; // 백엔드에서 소셜 로그인 URL을 생성하고 리다이렉트
 };
 
-export const handleSocialLoginCallback = async (provider: "naver" | "kakao", code: string, state?: string) => {
-  const response = await axios.post<LoginResponse>(`/api/auth/${provider}/callback`, { code, state });
+export const postSocialLoginCallback = async (provider: "naver" | "kakao", code: string, state?: string) => {
+  const response = await axios.post<LoginResponse>(`/oauth2/authorization/${provider}`, { code, state });
   return response.data;
 };
+// export const postSocialLoginCallback = async (provider: "naver" | "kakao", code: string, state?: string) => {
+//   const response = await axios.post<LoginResponse>(`/api/auth/${provider}/callback`, { code, state });
+//   return response.data;
+// };
 
 export const signUp = async (userData: SignUpData): Promise<void> => {
   await axios.post("/api/members/sign-up", userData);
